+++ libs/common/src/lib/interfaces/access.interface.ts
@@ -1,6 +1,6 @@
 export interface Access {
   alias?: string;
-  grantee: string;
+  grantee?: string;
   id: string;
-  type: 'PUBLIC' | 'RESTRICTED_VIEW';
+  type: 'PUBLIC' | 'PRIVATE' | 'RESTRICTED_VIEW';
 }